Fix all errors and make compatible with Android 2.2
Also got rid of the Arrow class to make way for using the touchscreen.
Added a really basic Player interface, the start of a HumanPlayer class,
a DrawableEntity base class that the board and counters extend, and got
it to the point that the blank activity runs on the emulator.